Modified folder structure, moved cursor info from displaycontroller to vimgrid
