+++ README.md
@@ -1,2 +1,41 @@
-# Lab2
-Fresh repository with main branch.
+# VimBeat
+
+## Dependencies
+
+### Runtime Dependencies
+- `konva` - 2D canvas library for animations and interactions
+
+### Development Dependencies
+- `typescript` - TypeScript compiler
+- `vite` - Build tool and dev server
+
+## Prerequisites
+
+- Node.js (v14 or higher)
+- npm (comes with Node.js)
+
+## Installation
+
+Install all dependencies:
+
+```bash
+npm install
+```
+
+This command automatically installs all runtime and development dependencies defined in `package.json`.
+
+## Running
+
+### Development Server
+
+```bash
+npm start
+```
+
+or
+
+```bash
+npm run dev
+```
+
+The application will be available at `http://localhost:5173`
