+++ src/VimGrid.ts
@@ -1,4 +1,3 @@
-
 // Represents a single cell in the Vim grid with character and optional highlight
 // The highlight is how the character should be displayed (e.g., color, style)
 export type Cell = { ch: string; hl?: string };
@@ -7,6 +6,7 @@ export default class VimGrid {
     readonly numRows: number;
     readonly numCols: number;
     private grid: Cell[][];
+    private cursor: { row: number; col: number };
 
     constructor(numRows: number, numCols: number, Cell: Cell = { ch: ' ' }) {
         this.numRows = numRows;
@@ -14,14 +14,18 @@ export default class VimGrid {
         this.grid = Array.from({ length: numRows }, () =>
             Array.from({ length: numCols }, () => ({ ...Cell }))
         );
+        this.cursor = { row: 0, col: 0 };
     }
 
     // Creates a VimGrid from a 2d character array
-    static createGridFromText(lines: string[], cols?: number) {
+    static createGridFromText(lines: string[], cols?: number, initialCursor?: { row: number; col: number }) {
         const width = cols ?? Math.max(1, ...lines.map((l) => l.length));
         const height = Math.max(lines.length, 1);
         const buf = new VimGrid(height, width);
         lines.forEach((line, r) => { for (let c = 0; c < width; c++) buf.set(r, c, { ch: line[c] ?? " " }); });
+        if (initialCursor) {
+            buf.setCursor(initialCursor.row, initialCursor.col);
+        }
         return buf;
     }
 
@@ -50,5 +54,21 @@ export default class VimGrid {
         return this.grid;
     }
 
+    // Cursor methods
+    getCursor(): { row: number; col: number } {
+        return { ...this.cursor };
+    }
 
+    setCursor(row: number, col: number): void {
+        const r = Math.max(0, Math.min(this.numRows - 1, row));
+        const c = Math.max(0, Math.min(this.numCols - 1, col));
+        this.cursor = { row: r, col: c };
+    }
+
+    moveCursorBy(dr: number, dc: number): void {
+        const r = Math.max(0, Math.min(this.numRows - 1, this.cursor.row + dr));
+        const c = Math.max(0, Math.min(this.numCols - 1, this.cursor.col + dc));
+        this.cursor = { row: r, col: c };
+    }
   }
+
